refactor: useMainPage 네이밍 수정
<before>
- useMainPage

<after>
- useProductList
