+++ src/pages/MainPage/MainPage.tsx
@@ -1,9 +1,9 @@
 import ProductCardGrid from '../../components/ProductCardGrid/ProductCardGrid';
 import Spinner from '../../components/Spinner/Spinner';
-import useMainPage from './useMainPage';
+import useProductList from './useProductList';
 
 function MainPage() {
-  const { isLoading, productList } = useMainPage();
+  const { isLoading, productList } = useProductList();
 
   if (isLoading) return <Spinner />;
